+++ R/tarpuy_plex.R
@@ -326,13 +326,63 @@ logbook <- tibble(Date = c(rep(NA, 3), as.character.Date(start), rep(NA, 3))
 
 # budget ------------------------------------------------------------------
 
-budget <- tibble("Material/Service" = rep(NA, 5)
-                 , Unit = rep(NA, 5)
-                 , Price = rep(NA, 5)
-                 , Quantity = rep(NA, 5)
-                 , Total = rep(NA, 5)
-                 , Description = rep(NA, 5)
-                 )
+budget <- tibble(
+  Category = c("Supplies", "Supplies", "Supplies", "Materials", "Materials", 
+               "Equipment", "Tools", "Equipment", "Human Resources", 
+               "Human Resources", "Human Resources", "Transport and Logistics", 
+               "Transport and Logistics", "Services", "Services", "Other", 
+               "Other", "Other"),
+  Description = c("Seeds", "Fertilizers", "Phytosanitary products", 
+                  "Substrate (soil, sand, etc.)", "Laboratory materials (pipettes, Petri dishes, etc.)", 
+                  "Measuring instruments (sensors, scales, etc.)", "Field tools", 
+                  "Software or programs for data analysis", "Researchers", 
+                  "Laboratory technicians", "Field workers", "Equipment transport", 
+                  "Personnel transport", "Sequencing", "Article publication", 
+                  "Electricity", "General services", "Administrative expenses"),
+  Quantity = NA,
+  Unit = c(NA, NA, NA, "kg", NA, NA, NA, NA, "months", "months", "day", 
+           "trips", "trips", "samples", "articles", "months", NA, NA),
+  `Unit Cost` = NA,
+  `Total Cost` = NA,
+  `Technical Specifications` = c("Variety, purity, germination (%)", 
+                               "Chemical composition (NPK, etc.)", 
+                               "Active ingredient, concentration", 
+                               "Soil type, pH, electrical conductivity", 
+                               "Brand, capacity, material", 
+                               "Precision, measurement range", 
+                               "Type of tool, material", 
+                               "Software name, license, compatibility", 
+                               "Level of specialization, required hours", 
+                               "Specialization, certification", 
+                               "Experience, skills", 
+                               "Type of vehicle, distance", 
+                               "Number of people, distance", 
+                               "Type of sequencing, number of samples", 
+                               "Journal, impact level", 
+                               "Kilowatt per month", 
+                               "Service details", 
+                               "Details of administrative expenses"),
+  Justification = c("Essential for experimental cultivation", 
+                    "Provides essential nutrients for growth", 
+                    "Prevents pests and diseases", 
+                    "Necessary for seed germination", 
+                    "For precise measurements and laboratory cultivation", 
+                    "Measurement of environmental and growth variables", 
+                    "For crop management in the field", 
+                    "Processing and analysis of experimental data", 
+                    "Specialized personnel for design and analysis", 
+                    "Technical assistance for experiment management", 
+                    "Implementation and management of field experiments", 
+                    "Transport of necessary equipment to the experimental site", 
+                    "Transport of personnel to the experimental site", 
+                    "Genetic analysis of samples", 
+                    "Publication of experimental results", 
+                    "Power supply for laboratory and field equipment", 
+                    "Necessary for general experiment maintenance", 
+                    "Administration and management of the experimental project"),
+  `Reference` = "<link to info>"
+)
+
 
 # result ------------------------------------------------------------------